pyproject.toml
[project]
name = "my-microservice"
version = "0.1.0"
description = "FastAPI Microservice"
authors = [
    {name = "Tapesh",email = "tapeshchawle@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
